==> 03.integration/src/main/R/07.01.Allen_ref_cleanup.by.Johny.20240522.R <==
## Clean up Allen reference scRNA-seq data
### Zhaoning Wang 20240522

library(Seurat)
library(Signac)
library(ggplot2)
library(ggrepel)
library(ggridges)
library(harmony)
library(cowplot)
library(dplyr)
library(RColorBrewer)
library(S4Vectors)
library(sctransform)
library(reticulate)
library(SeuratData)
library(SeuratDisk)
library(anndata)
library(irlba)
library(SeuratObject)

setwd("/projects/ps-renlab2/zhw063/99.MouseBrainPairedTag/allen_brain_atlas/allen_ref_seurat_object/")

AMY.allen <- readRDS("./AMY.allen.rna.rds")
CPU.allen <- readRDS("./CPU.allen.rna.rds")
HYP.allen <- readRDS("./HYP.allen.rna.rds")
HIP.allen <- readRDS("./HIP.allen.rna.rds")
ERC.allen <- readRDS("./ERC.allen.rna.rds")
NAC.allen <- readRDS("./NAC.allen.rna.rds")
VTA.allen <- readRDS("./VTA.allen.rna.rds")
PFC.allen <- readRDS("./PFC.allen.rna.rds")

AMY.allen[["manualcheck"]] <- "discard"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "319 Astro-TE NN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "064 STR-PAL Chst9 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "026 NLOT Rho Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "114 COAa-PAA-MEA Barhl2 Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "062 STR D2 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "074 MEA-BST Lhx6 Sp9 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "113 MEA-COA-BMA Ccdc42 Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "326 OPC NN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "318 Astro-NT NN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "014 LA-BLA-BMA-PA Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "327 Oligo NN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "012 MEA Slc17a7 Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "073 MEA-BST Sox6 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "079 CEA-BST Six3 Cyp26b1 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "080 CEA-AAA-BST Six3 Sp9 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "082 CEA-BST Ebf1 Pdyn Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "053 Sst Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "076 MEA-BST Lhx6 Nfib Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "061 STR D1 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "057 NDB-SI-MA-STRv Lhx8 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "334 Microglia NN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "036 HPF CR Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "330 VLMC NN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "119 SI-MA-LPO-LHA Skor1 Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "325 CHOR NN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "046 Vip Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "333 Endo NN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "120 MEA Otp Foxp2 Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "054 STR Prox1 Lhx6 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "045 OB-STR-CTX Inh IMN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "063 STR D1 Sema5a Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "075 MEA-BST Lhx6 Nr2e1 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "047 Sncg Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "049 Lamp5 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "323 Ependymal NN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "052 Pvalb Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "090 BST-MPN Six3 Nrgn Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "085 SI-MPO-LPO Lhx8 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "056 Sst Chodl Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "066 NDB-SI-ant Prdm12 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "139 PH-LHA Foxb1 Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "058 PAL-STR Gaba-Chol" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "329 ABC NN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "121 MEA-BST Otp Zic2 Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "088 BST Tac2 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "084 BST-SI-AAA Six3 Slc22a3 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "065 IA Mgp Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "083 CEA-BST Rai14 Pdyn Crh Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "051 Pvalb chandelier Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "050 Lamp5 Lhx6 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "131 LHA-AHN-PVH Otp Trh Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "055 STR Lhx8 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "048 RHP-COA Ndnf Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "331 Peri NN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "081 ACB-BST-FS D1 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "059 GPe-SI Sox6 Cyp26b1 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "077 CEA-BST Gal Avp Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "335 BAM NN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "078 SI-MA-ACB Ebf1 Bnc2 Gaba" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "117 LHA Barhl2 Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "118 ADP-MPO Trp73 Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "013 COAp Grxcr2 Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "332 SMC NN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "037 DG Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "322 Tanycyte NN" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "122 LHA-MEA Otp Glut" ] <- "select"
AMY.allen@meta.data$manualcheck[AMY.allen@meta.data$subclass_id_label == "338 Lymphoid NN" ] <- "select"


CPU.allen[["manualcheck"]] <- "discard"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "062 STR D2 Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "061 STR D1 Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "319 Astro-TE NN" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "327 Oligo NN" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "045 OB-STR-CTX Inh IMN" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "326 OPC NN" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "063 STR D1 Sema5a Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "334 Microglia NN" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "323 Ependymal NN" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "054 STR Prox1 Lhx6 Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "064 STR-PAL Chst9 Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "056 Sst Chodl Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "333 Endo NN" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "058 PAL-STR Gaba-Chol" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "055 STR Lhx8 Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "030 L6 CT CTX Glut" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "004 L6 IT CTX Glut" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "318 Astro-NT NN" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "052 Pvalb Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "330 VLMC NN" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "331 Peri NN" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "053 Sst Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "029 L6b CTX Glut" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "041 OB-in Frmd7 Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "049 Lamp5 Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "002 IT EP-CLA Glut" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "335 BAM NN" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "046 Vip Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "047 Sncg Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "079 CEA-BST Six3 Cyp26b1 Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "014 LA-BLA-BMA-PA Glut" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "080 CEA-AAA-BST Six3 Sp9 Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "059 GPe-SI Sox6 Cyp26b1 Gaba" ] <- "select"
CPU.allen@meta.data$manualcheck[CPU.allen@meta.data$subclass_id_label == "332 SMC NN" ] <- "select"


ERC.allen[["manualcheck"]] <- "discard"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "008 L2/3 IT ENT Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "319 Astro-TE NN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "028 L6b/CT ENT Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "019 L2/3 IT PPP Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "326 OPC NN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "009 L2/3 IT PIR-ENTl Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "330 VLMC NN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "011 L2 IT ENT-po Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "052 Pvalb Gaba" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "327 Oligo NN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "003 L5/6 IT TPE-ENT Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "053 Sst Gaba" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "036 HPF CR Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "329 ABC NN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "030 L6 CT CTX Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "045 OB-STR-CTX Inh IMN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "333 Endo NN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "023 SUB-ProS Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "031 CT SUB Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "046 Vip Gaba" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "049 Lamp5 Gaba" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "334 Microglia NN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "047 Sncg Gaba" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "015 ENTmv-PA-COAp Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "007 L2/3 IT CTX Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "318 Astro-NT NN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "002 IT EP-CLA Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "005 L5 IT CTX Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "323 Ependymal NN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "022 L5 ET CTX Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "016 CA1-ProS Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "001 CLA-EPd-CTX Car3 Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "004 L6 IT CTX Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "027 L6b EPd Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "034 NP PPP Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "050 Lamp5 Lhx6 Gaba" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "013 COAp Grxcr2 Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "029 L6b CTX Glut" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "038 DG-PIR Ex IMN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "051 Pvalb chandelier Gaba" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "048 RHP-COA Ndnf Gaba" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "331 Peri NN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "056 Sst Chodl Gaba" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "335 BAM NN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "325 CHOR NN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "332 SMC NN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "338 Lymphoid NN" ] <- "select"
ERC.allen@meta.data$manualcheck[ERC.allen@meta.data$subclass_id_label == "337 DC NN" ] <- "select"

HIP.allen[["manualcheck"]] <- "discard"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "037 DG Glut" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "016 CA1-ProS Glut" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "319 Astro-TE NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "017 CA3 Glut" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "327 Oligo NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "326 OPC NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "325 CHOR NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "038 DG-PIR Ex IMN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "323 Ependymal NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "023 SUB-ProS Glut" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "025 CA2-FC-IG Glut" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "053 Sst Gaba" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "330 VLMC NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "334 Microglia NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "036 HPF CR Glut" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "050 Lamp5 Lhx6 Gaba" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "333 Endo NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "052 Pvalb Gaba" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "318 Astro-NT NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "046 Vip Gaba" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "048 RHP-COA Ndnf Gaba" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "047 Sncg Gaba" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "031 CT SUB Glut" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "033 NP SUB Glut" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "051 Pvalb chandelier Gaba" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "045 OB-STR-CTX Inh IMN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "049 Lamp5 Gaba" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "329 ABC NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "015 ENTmv-PA-COAp Glut" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "335 BAM NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "331 Peri NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "322 Tanycyte NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "332 SMC NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "056 Sst Chodl Gaba" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "321 Astroependymal NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "338 Lymphoid NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "337 DC NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "336 Monocytes NN" ] <- "select"
HIP.allen@meta.data$manualcheck[HIP.allen@meta.data$subclass_id_label == "145 MH Tac2 Glut" ] <- "select"


HYP.allen[["manualcheck"]] <- "discard"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "318 Astro-NT NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "101 ZI Pax6 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "327 Oligo NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "127 DMH-LHA Vgll2 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "129 VMH Nr5a1 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "326 OPC NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "323 Ependymal NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "106 PVpo-VMPO-MPN Hmx2 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "098 AHN-SBPV-PVHd Pdrm12 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "319 Astro-TE NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "133 PVH-SO-PVa Otp Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "090 BST-MPN Six3 Nrgn Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "125 DMH Hmx2 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "107 DMH Hmx2 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "097 PVHd-SBPV Six3 Prox1 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "132 AHN-RCH-LHA Otp Fezf1 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "076 MEA-BST Lhx6 Nfib Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "100 AHN Onecut3 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "074 MEA-BST Lhx6 Sp9 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "152 RE-Xi Nox4 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "094 SCH Six6 Cdc14a Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "104 TU-ARH Otp Six6 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "138 PH Pitx2 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "322 Tanycyte NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "139 PH-LHA Foxb1 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "131 LHA-AHN-PVH Otp Trh Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "091 ARH-PVi Six6 Dopa-Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "093 RT-ZI Gnb3 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "099 SBPV-PVa Six6 Satb2 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "334 Microglia NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "085 SI-MPO-LPO Lhx8 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "073 MEA-BST Sox6 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "089 PVR Six3 Sox3 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "135 STN-PSTN Pitx2 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "109 LGv-ZI Otx2 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "126 ARH-PVp Tbx3 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "103 PVHd-DMH Lhx6 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "117 LHA Barhl2 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "102 DMH-LHA Gsx1 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "333 Endo NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "119 SI-MA-LPO-LHA Skor1 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "330 VLMC NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "136 PMv-TMv Pitx2 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "151 TH Prkcd Grin2c Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "124 MPN-MPO-PVpo Hmx2 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "118 ADP-MPO Trp73 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "128 VMH Fezf1 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "086 MPO-ADP Lhx8 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "130 LHA Pmch Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "134 PH-ant-LHA Otp Bsx Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "057 NDB-SI-MA-STRv Lhx8 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "080 CEA-AAA-BST Six3 Sp9 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "123 DMH Nkx2-4 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "114 COAa-PAA-MEA Barhl2 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "202 PRT Tcf7l2 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "088 BST Tac2 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "084 BST-SI-AAA Six3 Slc22a3 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "141 PH-SUM Foxa1 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "077 CEA-BST Gal Avp Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "112 GPi Tbr1 Cngb3 Gaba-Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "142 HY Gnrh1 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "329 ABC NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "012 MEA Slc17a7 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "095 DMH Prdm13 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "105 TMd-DMH Foxd2 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "046 Vip Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "331 Peri NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "113 MEA-COA-BMA Ccdc42 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "064 STR-PAL Chst9 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "140 PMd-LHA Foxb1 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "092 TMv-PMv Tbx3 Hist-Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "058 PAL-STR Gaba-Chol" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "062 STR D2 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "082 CEA-BST Ebf1 Pdyn Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "137 PH-an Pitx2 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "055 STR Lhx8 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "079 CEA-BST Six3 Cyp26b1 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "078 SI-MA-ACB Ebf1 Bnc2 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "168 SPA-SPFm-SPFp-POL-PIL-PoT Sp9 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "087 MPN-MPO-LPO Lhx6 Zfhx3 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "335 BAM NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "054 STR Prox1 Lhx6 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "061 STR D1 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "203 LGv-SPFp-SPFm Nkx2-2 Tcf7l2 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "215 SNc-VTA-RAmb Foxa1 Dopa" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "053 Sst Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "059 GPe-SI Sox6 Cyp26b1 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "081 ACB-BST-FS D1 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "056 Sst Chodl Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "075 MEA-BST Lhx6 Nr2e1 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "063 STR D1 Sema5a Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "121 MEA-BST Otp Zic2 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "066 NDB-SI-ant Prdm12 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "083 CEA-BST Rai14 Pdyn Crh Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "060 OT D3 Folh1 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "120 MEA Otp Foxp2 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "332 SMC NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "159 IF-RL-CLI-PAG Foxa1 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "213 SCsg Gabrr2 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "325 CHOR NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "143 MM-ant Foxb1 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "115 MS-SF Bsx Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "144 MM Foxb1 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "116 AVPV-MEPO-SFO Tbr1 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "108 ARH-PVp Tbx3 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "149 PVT-PT Ntrk1 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "096 PVHd Gsc Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "167 PRC-PAG Tcf7l2 Irx2 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "069 LSX Nkx2-1 Gaba" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "122 LHA-MEA Otp Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "156 MB-ant-ve Dmrta2 Glut" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "321 Astroependymal NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "338 Lymphoid NN" ] <- "select"
HYP.allen@meta.data$manualcheck[HYP.allen@meta.data$subclass_id_label == "337 DC NN" ] <- "select"


NAC.allen[["manualcheck"]] <- "discard"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "061 STR D1 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "062 STR D2 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "319 Astro-TE NN" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "063 STR D1 Sema5a Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "010 IT AON-TT-DP Glut" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "327 Oligo NN" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "064 STR-PAL Chst9 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "045 OB-STR-CTX Inh IMN" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "326 OPC NN" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "060 OT D3 Folh1 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "334 Microglia NN" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "333 Endo NN" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "056 Sst Chodl Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "054 STR Prox1 Lhx6 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "055 STR Lhx8 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "323 Ependymal NN" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "318 Astro-NT NN" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "058 PAL-STR Gaba-Chol" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "036 HPF CR Glut" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "030 L6 CT CTX Glut" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "053 Sst Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "081 ACB-BST-FS D1 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "052 Pvalb Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "080 CEA-AAA-BST Six3 Sp9 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "066 NDB-SI-ant Prdm12 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "029 L6b CTX Glut" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "057 NDB-SI-MA-STRv Lhx8 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "331 Peri NN" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "027 L6b EPd Glut" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "046 Vip Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "084 BST-SI-AAA Six3 Slc22a3 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "119 SI-MA-LPO-LHA Skor1 Glut" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "085 SI-MPO-LPO Lhx8 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "335 BAM NN" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "047 Sncg Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "330 VLMC NN" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "079 CEA-BST Six3 Cyp26b1 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "041 OB-in Frmd7 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "049 Lamp5 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "082 CEA-BST Ebf1 Pdyn Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "078 SI-MA-ACB Ebf1 Bnc2 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "039 OB Meis2 Thsd7b Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "004 L6 IT CTX Glut" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "059 GPe-SI Sox6 Cyp26b1 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "042 OB-out Frmd7 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "329 ABC NN" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "002 IT EP-CLA Glut" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "065 IA Mgp Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "332 SMC NN" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "083 CEA-BST Rai14 Pdyn Crh Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "044 OB Dopa-Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "070 LSX Prdm12 Slit2 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "040 OB Trdn Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "337 DC NN" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "035 OB Eomes Ms4a15 Glut" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "071 LSX Prdm12 Zeb2 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "067 LSX Sall3 Pax6 Gaba" ] <- "select"
NAC.allen@meta.data$manualcheck[NAC.allen@meta.data$subclass_id_label == "069 LSX Nkx2-1 Gaba" ] <- "select"

PFC.allen[["manualcheck"]] <- "discard"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "007 L2/3 IT CTX Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "006 L4/5 IT CTX Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "030 L6 CT CTX Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "319 Astro-TE NN" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "022 L5 ET CTX Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "005 L5 IT CTX Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "004 L6 IT CTX Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "052 Pvalb Gaba" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "032 L5 NP CTX Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "326 OPC NN" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "053 Sst Gaba" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "333 Endo NN" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "327 Oligo NN" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "330 VLMC NN" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "045 OB-STR-CTX Inh IMN" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "046 Vip Gaba" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "047 Sncg Gaba" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "049 Lamp5 Gaba" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "334 Microglia NN" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "329 ABC NN" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "029 L6b CTX Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "050 Lamp5 Lhx6 Gaba" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "051 Pvalb chandelier Gaba" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "331 Peri NN" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "056 Sst Chodl Gaba" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "003 L5/6 IT TPE-ENT Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "020 L2/3 IT RSP Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "010 IT AON-TT-DP Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "335 BAM NN" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "009 L2/3 IT PIR-ENTl Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "332 SMC NN" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "002 IT EP-CLA Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "021 L4 RSP-ACA Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "042 OB-out Frmd7 Gaba" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "041 OB-in Frmd7 Gaba" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "001 CLA-EPd-CTX Car3 Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "337 DC NN" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "044 OB Dopa-Gaba" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "338 Lymphoid NN" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "039 OB Meis2 Thsd7b Gaba" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "035 OB Eomes Ms4a15 Glut" ] <- "select"
PFC.allen@meta.data$manualcheck[PFC.allen@meta.data$subclass_id_label == "336 Monocytes NN" ] <- "select"


VTA.allen[["manualcheck"]] <- "discard"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "327 Oligo NN" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "318 Astro-NT NN" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "215 SNc-VTA-RAmb Foxa1 Dopa" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "135 STN-PSTN Pitx2 Glut" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "326 OPC NN" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "319 Astro-TE NN" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "197 SNr Six3 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "195 SNr-VTA Pax5 Npas1 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "139 PH-LHA Foxb1 Glut" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "203 LGv-SPFp-SPFm Nkx2-2 Tcf7l2 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "330 VLMC NN" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "334 Microglia NN" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "333 Endo NN" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "196 PAG-PPN Pax5 Sox21 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "093 RT-ZI Gnb3 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "166 MRN Pou3f1 C1ql4 Glut" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "214 IPN Otp Crisp1 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "191 PAG-MRN Rln3 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "335 BAM NN" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "159 IF-RL-CLI-PAG Foxa1 Glut" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "331 Peri NN" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "213 SCsg Gabrr2 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "202 PRT Tcf7l2 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "199 PAG-MRN-RN Foxa2 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "200 PAG-ND-PCG Onecut1 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "216 MB-MY Tph2 Glut-Sero" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "332 SMC NN" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "053 Sst Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "052 Pvalb Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "050 Lamp5 Lhx6 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "016 CA1-ProS Glut" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "047 Sncg Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "048 RHP-COA Ndnf Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "049 Lamp5 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "168 SPA-SPFm-SPFp-POL-PIL-PoT Sp9 Glut" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "051 Pvalb chandelier Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "156 MB-ant-ve Dmrta2 Glut" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "165 PAG-MRN Pou3f1 Glut" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "180 SCiw Pitx2 Glut" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "268 CS-PRNr-DR En1 Sox2 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "198 IC Six3 En2 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "231 IPN-LDT Vsx2 Nkx6-1 Glut" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "201 PAG-RN Nkx2-2 Otx1 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "338 Lymphoid NN" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "193 MRN-PPN-CUN Pax8 Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "267 CS-PRNr-PCG Tmem163 Otp Gaba" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "336 Monocytes NN" ] <- "select"
VTA.allen@meta.data$manualcheck[VTA.allen@meta.data$subclass_id_label == "337 DC NN" ] <- "select"


SaveSeuratRds(object = AMY.allen,file = "./20240522.AMY.clean.allen.rna.rds")
SaveSeuratRds(object = CPU.allen,file = "./20240522.CPU.clean.allen.rna.rds")
SaveSeuratRds(object = ERC.allen,file = "./20240522.ERC.clean.allen.rna.rds")
SaveSeuratRds(object = HIP.allen,file = "./20240522.HIP.clean.allen.rna.rds")
SaveSeuratRds(object = HYP.allen,file = "./20240522.HYP.clean.allen.rna.rds")
SaveSeuratRds(object = NAC.allen,file = "./20240522.NAC.clean.allen.rna.rds")
SaveSeuratRds(object = PFC.allen,file = "./20240522.PFC.clean.allen.rna.rds")
SaveSeuratRds(object = VTA.allen,file = "./20240522.VTA.clean.allen.rna.rds")
